Fix wrong type for GO_Type enum
Prior save games used 2 Bytes instead of 1 (while 1 would have been enough)
Fixes #1328

+++ libs/s25main/SerializedGameData.cpp
@@ -341,7 +341,7 @@ void SerializedGameData::PushObject_(const GameObject* go, const bool known)
 
     // Objekt nich bekannt? Dann Type-ID noch mit drauf
     if(!known)
-        PushEnum<uint8_t>(go->GetGOT());
+        PushEnum<uint16_t>(go->GetGOT());
 
     // Objekt serialisieren
     if(debugMode)
